requirements.txt
python-dotenv
matplotlib
# Core Data Processing
pandas>=2.3.2
numpy>=2.3.3
pytz
python-dateutil

# Webex/Communication
webexpythonsdk==2.0.4 # pinned to 2.0.4 for compatibility with webex_bot 1.0.8
webexteamssdk # deprecated
webex_bot==1.0.8 # the framework handles the WebSocket connection setup

# Web Frameworks
Flask
waitress
altair

# Microsoft Teams Bot Framework
botbuilder-core
botbuilder-schema
botframework-connector

# HTTP/Network
requests
aiohttp
urllib3
websockets>=11.0.3,<15.0  # webex_bot 1.0.8 requires 11.0.3, but works with 11.0.3-14.2, v15+ has breaking async changes (requires enhanced_websocket_client.py compatibility patch for 12.0+). Install 14.2 specifically after installing requirements.
aiohttp-retry

# Security/Authentication - Always use latest for security
PyJWT
cryptography
certifi

# Azure/Cloud
azure-core
azure-devops
msrest

# Data Processing/Analysis
openpyxl
tabulate
beautifulsoup4
pyarrow
python-docx
unstructured
networkx
msoffcrypto-tool

# Image Processing
pillow
opencv-python-headless
pytesseract
Cartopy

# Security Tools
crowdstrike-falconpy
demisto-py>=3.2.21  # Official Palo Alto XSOAR Python SDK
dnstwist  # Domain typosquatting detection
python-whois  # WHOIS lookup library

# Task Scheduling/Utilities
schedule
tqdm
fasteners
tenacity
filelock
config

# AI/LangChain - Use latest for features and fixes
langchain>=1.1.3
langchain-community>=0.4.1
langchain-core>=1.1.3
langchain-ollama>=1.0.0
ollama>=0.6.1  # Ollama Python client (used by langchain-ollama)
langsmith
rank_bm25
uuid-utils>=0.12.0  # Required by langchain-core 1.1.3+

# Local RAG Components
sentence-transformers>=5.1.0
chromadb>=1.0.21
pypdf>=6.0.0

# Development/Testing
pytest
types-pytz
jsonschema
pydantic-settings
typing_extensions

texttable
